+++ src/pages/sign/SignUpForm.tsx
@@ -16,13 +16,6 @@ import {
   FormLabel,
   FormMessage,
 } from "@/components/ui/form";
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from "@/components/ui/select";
 
 import { ToastAction } from "@/components/ui/toast";
 import { useToast } from "@/components/ui/use-toast";
@@ -36,7 +29,6 @@ import { zodResolver } from "@hookform/resolvers/zod";
 import { z } from "zod";
 import { useEffect, useState } from "react";
 import { ArrowRight } from "lucide-react";
-import { RadioGroup, RadioGroupItem } from "@/components/ui/radio-group";
 import { useRouter } from "next/router";
 import { auth, db } from "../api/firebaseSDK";
 import { addDoc, collection, serverTimestamp } from "firebase/firestore";
@@ -45,14 +37,14 @@ import {
   onAuthStateChanged,
 } from "firebase/auth";
 import { UserType, addrType } from "@/types/type";
-import { randomUUID } from "crypto";
+import SignUpFirstStep from "./SignUpFirstStep";
 
 type RegisterType = z.infer<typeof registerSchema>;
 
 const SignUpForm = () => {
   const [step, setStep] = useState<number>(0);
   const [kakaoAddr, setKakaoAddr] = useState<string>("");
-  const [postCode, setPostCode] = useState<string>("");
+  // const [postCode, setPostCode] = useState<string>("");
   const { toast } = useToast();
   const router = useRouter();
 
@@ -159,7 +151,7 @@ const SignUpForm = () => {
       });
   };
   // 다음 단계 핸들러
-  const nextLevelHandler = async () => {
+  const nextStepHandler = async () => {
     // 카카오 주소 유효성 검사
     if (kakaoAddr === "") {
       toast({
@@ -199,24 +191,6 @@ const SignUpForm = () => {
     });
   };
 
-  const kakaoAddrModal = () => {
-    if (!window.daum) {
-      console.error("window.daum is not defined");
-      return;
-    }
-    new window.daum.Postcode({
-      oncomplete: function (data: addrType) {
-        let selectedAddress = data.address;
-        // 지번주소 선택 시 지번주소 저장
-        if (data.userSelectedType === "J") {
-          selectedAddress = data.jibunAddress;
-        }
-        setKakaoAddr(selectedAddress);
-        setPostCode(data.zonecode);
-      },
-    }).open();
-  };
-
   return (
     <Layout>
       <Card className={cn("w-full")}>
@@ -230,247 +204,13 @@ const SignUpForm = () => {
               onSubmit={form.handleSubmit(onSubmit)}
               className="relative overflow-x-hidden w-full"
             >
-              <motion.div
-                className={cn("flex justify-between gap-10")}
-                animate={{ translateX: `${step * -100}%` }}
-                transition={{ ease: "easeInOut" }}
-              >
-                <div className="w-[250px]">
-                  <FormField
-                    control={form.control}
-                    name="username"
-                    render={({ field }) => (
-                      <FormItem>
-                        <FormLabel>이름</FormLabel>
-                        <FormControl>
-                          <Input placeholder="이름을 입력하세요" {...field} />
-                        </FormControl>
-                        <FormMessage />
-                      </FormItem>
-                    )}
-                  />
-                  <FormField
-                    control={form.control}
-                    name="email"
-                    render={({ field }) => (
-                      <FormItem>
-                        <FormLabel>이메일</FormLabel>
-                        <FormControl>
-                          <Input placeholder="이메일을 입력하세요" {...field} />
-                        </FormControl>
-                        <FormMessage />
-                      </FormItem>
-                    )}
-                  />
-                  <FormField
-                    control={form.control}
-                    name="phone"
-                    render={({ field }) => (
-                      <FormItem>
-                        <FormLabel>연락처</FormLabel>
-                        <FormControl>
-                          <Input placeholder="010-XXXX-XXXX" {...field} />
-                        </FormControl>
-                        <FormMessage />
-                      </FormItem>
-                    )}
-                  />
-                  <FormField
-                    control={form.control}
-                    name="gender"
-                    render={({ field }) => (
-                      <FormItem>
-                        <FormLabel>성별</FormLabel>
-                        <FormControl>
-                          <RadioGroup
-                            onValueChange={field.onChange}
-                            defaultValue={field.value}
-                            className="flex flex-col"
-                          >
-                            <div className="flex gap-1">
-                              <FormControl>
-                                <RadioGroupItem value="남성" id="gender-male" />
-                              </FormControl>
-                              <FormLabel
-                                className="font-normal"
-                                htmlFor="gender-male"
-                              >
-                                남
-                              </FormLabel>
-
-                              <FormControl>
-                                <RadioGroupItem
-                                  value="여성"
-                                  id="gender-female"
-                                />
-                              </FormControl>
-                              <FormLabel
-                                className="font-normal"
-                                htmlFor="gender-female"
-                              >
-                                녀
-                              </FormLabel>
-                            </div>
-                          </RadioGroup>
-                        </FormControl>
-                        <FormMessage />
-                      </FormItem>
-                    )}
-                  />
-                  <div className="flex gap-3">
-                    <FormField
-                      control={form.control}
-                      name="birthYear"
-                      render={({ field }) => (
-                        <FormItem>
-                          <FormLabel>년</FormLabel>
-                          <Select
-                            onValueChange={field.onChange}
-                            defaultValue={field.value}
-                          >
-                            <FormControl>
-                              <SelectTrigger>
-                                <SelectValue placeholder="년" />
-                              </SelectTrigger>
-                            </FormControl>
-                            <SelectContent className="min-w-[20px]">
-                              {Array.from(
-                                { length: 114 },
-                                (_, i) => 2024 - i
-                              ).map((year) => (
-                                <SelectItem key={year} value={String(year)}>
-                                  {year}
-                                </SelectItem>
-                              ))}
-                            </SelectContent>
-                          </Select>
-                          <FormMessage />
-                        </FormItem>
-                      )}
-                    />
-                    <FormField
-                      control={form.control}
-                      name="birthMonth"
-                      render={({ field }) => (
-                        <FormItem>
-                          <FormLabel>월</FormLabel>
-                          <Select
-                            onValueChange={field.onChange}
-                            defaultValue={field.value}
-                          >
-                            <FormControl>
-                              <SelectTrigger>
-                                <SelectValue placeholder="월" />
-                              </SelectTrigger>
-                            </FormControl>
-                            <SelectContent className="min-w-[10px]">
-                              {Array.from({ length: 12 }, (_, i) => i + 1).map(
-                                (month) => (
-                                  <SelectItem key={month} value={String(month)}>
-                                    {String(month)}
-                                  </SelectItem>
-                                )
-                              )}
-                            </SelectContent>
-                          </Select>
-                          <FormMessage />
-                        </FormItem>
-                      )}
-                    />
-                    <FormField
-                      control={form.control}
-                      name="birthDay"
-                      render={({ field }) => (
-                        <FormItem>
-                          <FormLabel>일</FormLabel>
-                          <Select
-                            onValueChange={field.onChange}
-                            defaultValue={field.value}
-                          >
-                            <FormControl>
-                              <SelectTrigger>
-                                <SelectValue placeholder="일" />
-                              </SelectTrigger>
-                            </FormControl>
-                            <SelectContent className="min-w-[10px]">
-                              {Array.from({ length: 31 }, (_, i) => i + 1).map(
-                                (day) => (
-                                  <SelectItem key={day} value={String(day)}>
-                                    {day}
-                                  </SelectItem>
-                                )
-                              )}
-                            </SelectContent>
-                          </Select>
-                          <FormMessage />
-                        </FormItem>
-                      )}
-                    />
-                  </div>
-                </div>
-                <div className="mt-0">
-                  {/* 주소 입력란 */}
-                  <FormField
-                    control={form.control}
-                    name="detailAddr"
-                    render={({ field }) => (
-                      <FormItem>
-                        <FormLabel>주소</FormLabel>
-                        <FormControl>
-                          <div className="flex flex-col gap-3">
-                            <Input
-                              type="text"
-                              value={kakaoAddr}
-                              onClick={kakaoAddrModal}
-                              className="min-w-[300px]"
-                              readOnly
-                            />
-                            <Input
-                              type="text"
-                              value={postCode}
-                              onClick={kakaoAddrModal}
-                              readOnly
-                            />
-                            <Input
-                              id="addrDetail"
-                              placeholder="상세주소를 입력해 주세요."
-                              {...field}
-                            />
-                            <Button type="button" onClick={kakaoAddrModal}>
-                              주소 찾기
-                            </Button>
-                          </div>
-                        </FormControl>
-                        <FormMessage />
-                      </FormItem>
-                    )}
-                  />
-                  <FormField
-                    control={form.control}
-                    name="role"
-                    render={({ field }) => (
-                      <FormItem className="w-[200px]">
-                        <FormLabel>역할</FormLabel>
-                        <Select
-                          onValueChange={field.onChange}
-                          defaultValue={field.value}
-                        >
-                          <FormControl>
-                            <SelectTrigger>
-                              <SelectValue placeholder="역할을 선택해주세요" />
-                            </SelectTrigger>
-                          </FormControl>
-                          <SelectContent>
-                            <SelectItem value="admin">관리자</SelectItem>
-                            <SelectItem value="user">사용자</SelectItem>
-                          </SelectContent>
-                        </Select>
-                        <FormMessage />
-                      </FormItem>
-                    )}
-                  />
-                </div>
-              </motion.div>
+              {/* 첫번째 Step */}
+              <SignUpFirstStep
+                form={form}
+                kakaoAddr={kakaoAddr}
+                setKakaoAddr={setKakaoAddr}
+                step={step}
+              />
               <motion.div
                 className={cn("absolute top-0 left-0 right-0")}
                 animate={{ translateX: `${(1 - step) * 100}%` }}
@@ -514,7 +254,7 @@ const SignUpForm = () => {
                 <Button
                   type="button"
                   className={cn({ hidden: step === 1 })}
-                  onClick={nextLevelHandler}
+                  onClick={nextStepHandler}
                 >
                   다음 단계로
                   <ArrowRight className="w-4 h-4 ml-2" />
